nfccon/requirements.txt
Django==1.4.2
South==0.7.6
argparse==1.2.1
django-extensions==1.0.1
django-registration==0.8
djangorestframework==2.1.8
ipython==0.13.1
wsgiref==0.1.2
